+++ docs/_pages/releases.md
@@ -10,6 +10,7 @@ sidebar:
 
 ### What's new
 * Added support for MSTest v4 - [#312](https://github.com/awesomeassertions/awesomeassertions/issues/312)
+* Add `ExcludingMembersNamed` to `BeEquivalentTo` which enables excluding members with specified name - [#183](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/183)
 
 ### Improvements
 * Improve tabulator handling in string difference visualization - [#313](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/313)